KURSP-943-refactor some structure of folders and add course menu to announcements page
