+++ src/js/modules/menu.js
@@ -8,7 +8,7 @@ import { hrefQueryString } from '../settingsRoot.js';
 import i18n from '../i18n.js';
 import login from './login.js';
 import multilanguage from '../3party/multilanguage.js';
-import { renderCourseModules } from "../../vue/components/course-page-left-side-view/index"
+import { renderCourseModules } from "../../vue/pages/course-page/left-menu"
 import settings from '../settings.js';
 import usermenu from '../../templates/modules/usermenu.hbs';
 import util from './util.js'
