+++ src/vue/pages/course-page/left-menu/index.js
